% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.GSCIMC.R
\name{print.summary.GSCIMC}
\alias{print.summary.GSCIMC}
\title{Print Method for Summary of GSCIMC Models}
\usage{
\method{print}{summary.GSCIMC}(x, ...)
}
\arguments{
\item{x}{An object of class 'summary.GSCIMC'.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Print a summary for the specified Generalized Spatial Autoregresive Model class.
}
\description{
This method prints a formatted summary of a 'summary.GSCIMC' object,
including details of the model coefficients, rho, dispersion, and other statistics.
}
