% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.GEESAR.R
\name{print.summary.SARARgamlss}
\alias{print.summary.SARARgamlss}
\title{Print Method for Summary of SARARgamlss Models}
\usage{
\method{print}{summary.SARARgamlss}(x, ...)
}
\arguments{
\item{x}{An object of class 'summary.SARARgamlss'.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Print a summary for the specified GAMLSS model.
}
\description{
This method prints a formatted summary of a 'summary.SARARgamlss' object,
including details of the GAMLSS model, spatial parameters (rho and lambda), 
and Wald tests.
}
