% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SARARgamlss.R
\name{SARARgamlss}
\alias{SARARgamlss}
\title{Spatial autoregresive models with joint modelling of mean and variance}
\usage{
SARARgamlss(
  formula = formula(data),
  sigma.formula = ~1,
  W1 = diag(0, nrow(data)),
  W2 = diag(0, nrow(data)),
  data,
  tol = 1e-04,
  maxiter = 20,
  type = c("SAR", "SARAR", "SEM"),
  weights = NULL,
  ...
)
}
\description{
Function to calculate a SARAR, SEM or SAR model with explicit model using 
GAMLSS
}
