% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.GEESAR.R
\name{summary.GEESAR}
\alias{summary.GEESAR}
\title{Summary Method for GEESAR Objects}
\usage{
\method{summary}{GEESAR}(object, ...)
}
\arguments{
\item{object}{An object of class `"glmgee"`, generated by the GEESAR function.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A list of class `"summary.GEESAR"` containing:
  \item{coefficients}{A matrix of the estimated coefficients, standard errors, 
  z-values, and p-values.}
  \item{rho}{Estimated value of the correlation parameter.}
  \item{varrho}{Estimated variance of the correlation parameter rho.}
  \item{dispersion}{Estimated dispersion parameter of the model.}
  \item{df.residual}{Residual degrees of freedom.}
  \item{call}{The call used to generate the GEESAR model.}
  \item{family}{The family of the model.}
  \item{logLik}{The log-likelihood of the fitted model.}
  \item{CIC}{The Conditional Information Criterion for model fit.}
  \item{RJC}{The Jackknife-Covariance Ratio of the model.}
}
\description{
This function provides a summary of a GEESAR model object, which is typically 
a result of a generalized estimating equation (GEE) analysis using the GEESAR 
package. It returns a detailed summary of the model, including estimated coefficients, 
standard errors, z-statistics, p-values, rho (correlation parameter), 
variance of rho, and additional model statistics such as log-likelihood, 
conditional information criterion (CIC), and jackknife-covariance ratio (RJC).
}
\examples{
\dontrun{
  # Assuming you have fitted a GEESAR model
  out_ <- GEESAR(...)
  summary(out_)
}

}
\seealso{
\link[GEESAR]{GEESAR} for model fitting function.
}
