% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.GEESAR.R
\name{summary}
\alias{summary}
\title{Generic Summary Method for GEESAR}
\usage{
summary(object, ...)
}
\arguments{
\item{object}{An object to summarize.}

\item{...}{Additional arguments passed to the method.}
}
\description{
Defines a generic method for summarizing GEESAR model objects.
}
