% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.GEESAR.R
\name{print.summary.GEESAR}
\alias{print.summary.GEESAR}
\title{Print Method for Summary of GEESAR Models}
\usage{
\method{print}{summary.GEESAR}(x, ...)
}
\arguments{
\item{x}{An object of class 'summary.GEESAR'.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Print a summary for the specified Generalized Spatial Autoregresive Model class.
}
\description{
This method prints a formatted summary of a 'summary.GEESAR' object,
including details of the model coefficients, rho, dispersion, and other statistics.
}
