% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.GEESAR.R
\name{print.summary.GEESAR}
\alias{print.summary.GEESAR}
\title{Print Method for Summary of GEESAR Objects}
\usage{
\method{print}{summary.GEESAR}(x, ...)
}
\arguments{
\item{x}{An object of class `"summary.GEESAR"`.}

\item{...}{Additional arguments (currently unused).}
}
\description{
Prints a formatted summary of a GEESAR model.
}
