% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.GEESAR.R
\name{summary.GEESAR}
\alias{summary.GEESAR}
\title{Summary Method for GEESAR Objects}
\usage{
\method{summary}{GEESAR}(object, ...)
}
\arguments{
\item{object}{An object of class `"GEESAR"`, generated by the `GEESAR()` function.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A list of class `"summary.GEESAR"`.
}
\description{
Provides a summary of a GEESAR model object.
}
